pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "NSTstats"
version = "0.1.0"
description = "NSTstats project"
requires-python = ">=3.10"
# Add your other runtime dependencies below as needed
# Example: "pandas>=2.0", "requests>=2.31"
dependencies = [
  # Core data + HTTP
  "requests>=2.31",
  "python-dotenv>=1.0",

  # Parsing
  "beautifulsoup4>=4.12",
  "lxml>=4.9",

  # Data science
  "pandas>=2.0",
  "numpy>=1.24",

  # Yahoo OAuth client
  "yahoo-oauth>=2.0",

  # DB + migrations
  "SQLAlchemy>=2.0",
  "alembic>=1.13",
  # Optional: required when using Postgres (ignored for SQLite). Safe to include.
  "psycopg2-binary>=2.9",

  # Local package dependency (keep last so it's easy to spot)
  "nhl-schedule @ file:///C:/Users/soluk/PycharmProjects/NHLschedule",
]

[tool.setuptools]
# Automatically find packages (app, adapters, diagnostics, etc.) in the project root
packages = { find = {} }

[project.scripts]
# Optional CLI entry point; lets you run `nststats` after installation
nststats = "app.cli:main"